refacor delete and put routes w auhthentication

+++ spark-api/routes/ideas.js
@@ -45,19 +45,27 @@ router.post("/", async (req, res) => {
 //update and idea
 router.put("/:id", async (req, res) => {
   try {
-    const updatedIdea = await Idea.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          text: req.body.text,
-          tag: req.body.tag,
+    const idea = await Idea.findById(req.params.id);
+    if (idea.username === req.body.username) {
+      const updatedIdea = await Idea.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            text: req.body.text,
+            tag: req.body.tag,
+          },
         },
-      },
-      {
-        new: true,
-      }
-    );
-    res.json({ success: true, data: updatedIdea });
+        {
+          new: true,
+        }
+      );
+      return res.json({ success: true, data: updatedIdea });
+    }
+    //usernames does not match
+    res.status(403).json({
+      success: false,
+      error: "Not authorized to update this resource",
+    });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: "something went wrong" });
@@ -67,8 +75,15 @@ router.put("/:id", async (req, res) => {
 //delete an idea
 router.delete("/:id", async (req, res) => {
   try {
-    await Idea.findByIdAndDelete(req.params.id);
-    res.json({ success: true, data: {} });
+    //get idea and validate with username
+    const idea = await Idea.findById(req.params.id);
+    if (idea.username === req.body.username) {
+      await Idea.findByIdAndDelete(req.params.id);
+      return res.json({ success: true, data: {} });
+    }
+
+    //usernames do not match
+    res.status(403).json({ success: false, error: "Not authorized to delete" });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: "something went wrong" });
